code/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

cellpose==1.0.2
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.3.0
cycler==0.12.1
fastremap==1.19.0
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
idna==3.4
ImageIO==2.37.2
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy-loader==0.5
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==26.2
pandas==2.3.3
pillow==11.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.28.1
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.14.0
tifffile==2024.8.30
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.67.3
triton==2.1.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==1.26.13
zipp==3.23.1
